+++ src/neural_network/Wait.cpp
@@ -71,41 +71,41 @@ bool Wait::isOver(int epochs, float accuracy, float mae, int duration) const
 Wait snn::operator""_ep(unsigned long long value)
 {
     Wait res;
-    res.epochs = value;
+    res.epochs = (int)value;
     return res;
 }
 
 Wait snn::operator""_acc(long double value)
 {
     Wait res;
-    res.accuracy = value;
+    res.accuracy = (float)value;
     return res;
 }
 
 Wait snn::operator""_mae(long double value)
 {
     Wait res;
-    res.mae = value;
+    res.mae = (float)value;
     return res;
 }
 
 Wait snn::operator""_ms(unsigned long long value)
 {
     Wait res;
-    res.duration = value;
+    res.duration = (int)value;
     return res;
 }
 
 Wait snn::operator""_s(unsigned long long value)
 {
     Wait res;
-    res.duration = value * 1000;
+    res.duration = (int)value * 1000;
     return res;
 }
 
 Wait snn::operator""_min(unsigned long long value)
 {
     Wait res;
-    res.accuracy = value * 1000 * 60;
+    res.accuracy = (int)value * 1000 * 60;
     return res;
 }
